gl_common: signal to GL backend whether we are probing
This is supposed to reduce the amount of useless error messages shown
during initialization of vo_opengl. If multiple backends are compiled,
usually only one of them will work. For example, on Linux both X and
Wayland backends can be compiled, but usually either Wayland or X is
running. Then, if Wayland is not running, but X is, trying to initialize
the Wayland backend should not spam the terminal with error messages.

Code using this should probably raise the log level for error messages
from MSGL_ERR to MSGL_V. Only error messages that are typically printed
in these situations should be changed. For example, if Wayland is
available, but some other part of initialization fails, the error
messages should probably be visible.